Do 4.18 - 4.22.
Rook 1,5h

+++ controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
+const jwt = require('jsonwebtoken');
 
 blogsRouter.get('/', async (req, res) => {
   const blogs=await Blog
@@ -16,6 +17,7 @@ blogsRouter.get('/:id', async (req, res) => {
 
 blogsRouter.post('/', async (req, res) => {
   const body = req.body;
+  const user = await User.findById(req.user);
 
   if (!body.title) {
     return res.status(400).json({
@@ -29,14 +31,11 @@ blogsRouter.post('/', async (req, res) => {
     });
   }
 
-  //if (!body.user) {
-  //  return res.status(400).json({
-  //    error: 'Body content `user` missing'
-  //  });
-  //}
-
-  const user=await User.findById('61262b49067d229bf0f9f017');
-  console.log(user);
+  if (!user) {
+    return res.status(400).json({
+      error: 'no user with the given id ' + req.user
+    });
+  }
 
   const blog = new Blog({
     title: body.title,
@@ -55,8 +54,23 @@ blogsRouter.post('/', async (req, res) => {
 });
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id);
-  res.status(204).end();
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({error: 'token missing or invalid'});
+  }
+
+  const blog = await Blog.findById(req.params.id);
+  console.log(blog);
+  console.log(decodedToken.id);
+  console.log(blog.user.id);
+
+  if (req.user === blog.user.toString()) {
+    await Blog.findByIdAndRemove(req.params.id);
+    res.status(204).end();
+  } else {
+    res.status(401).end();
+  }
+
 });
 
 blogsRouter.put('/:id', async (req, res) => {
